+++ src/main/java/com/cinchapi/runway/Record.java
@@ -20,6 +20,7 @@ import groovy.lang.Sequence;
 
 import java.io.IOException;
 import java.io.Serializable;
+import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
@@ -358,6 +359,8 @@ public abstract class Record implements Comparable<Record> {
             @Nullable Map<String, Set<Object>> data, String prefix) {
         T record = (T) newDefaultInstance(clazz, connections);
         setInternalFieldValue("id", id, record);
+        setInternalFieldValue("waitingToBeDeleted", new LinkedHashSet<>(),
+                record);
         record.assign(runway);
         record.load(concourse, existing, data, prefix);
         record.onLoad();
@@ -639,6 +642,12 @@ public abstract class Record implements Comparable<Record> {
     private transient Map<String, Object> derived = null;
 
     /**
+     * Tracks dependent {@link Records} that are pending
+     * {@link #delete(Concourse) deletion} from {@link Concourse}.
+     */
+    private Set<Record> waitingToBeDeleted = new LinkedHashSet<>();
+
+    /**
      * Construct a new instance.
      */
     public Record() {
@@ -748,8 +757,8 @@ public abstract class Record implements Comparable<Record> {
     }
 
     /**
-     * Delete this {@link Record} from Concourse when the {@link #save()} method
-     * is called.
+     * Set this {@link Record} to be deleted from Concourse when the
+     * {@link #save()} method is called.
      */
     public void deleteOnSave() {
         deleted = true;
@@ -2059,7 +2068,96 @@ public abstract class Record implements Comparable<Record> {
      * @param concourse
      */
     private void deleteWithinTransaction(Concourse concourse) {
+        // Ensure any fields to which this Record must @CascadeDelete are
+        // deleted within this transaction
+        Set<Field> dependents = StaticAnalysis.instance()
+                .getAnnotatedFields(getClass(), CascadeDelete.class);
+        for (Field dependent : dependents) {
+            try {
+                Object value = dependent.get(this);
+                if(value instanceof Record) {
+                    ensureDeletion((Record) value);
+                }
+                else if(Sequences.isSequence(value)) {
+                    Sequences.forEach(value, item -> {
+                        if(item instanceof Record) {
+                            ensureDeletion((Record) item);
+                        }
+                    });
+                }
+            }
+            catch (ReflectiveOperationException e) {
+                throw CheckedExceptions.wrapAsRuntimeException(e);
+            }
+        }
+
+        // Check if there are any incoming links that used the @JoinDelete
+        // annotation to join in the deletion of this Record
+        Criteria potentialJoinDeletes = StaticAnalysis.instance()
+                .getJoinDeleteLookupCondition(this);
+        if(potentialJoinDeletes != null) {
+            for (Entry<Long, Set<Object>> entry : concourse
+                    .select(SECTION_KEY, potentialJoinDeletes).entrySet()) {
+                // It's necessary to load each of the Records (instead of
+                // directly clearing it in the database) in case any of them
+                // also have deletion hook annotations.
+                long id = entry.getKey();
+                String __ = (String) Iterables.getLast(entry.getValue());
+                Class<? extends Record> clazz = Reflection.getClassCasted(__);
+                Record record = db.load(clazz, id);
+                ensureDeletion(record);
+            }
+        }
+
+        // Check if there are any incoming links that used the @CaptureDelete
+        // annotation to remove references to this Record post deletion. Handle
+        // that business and save those records within this transaction
+        Criteria potentialCaptureDeletes = StaticAnalysis.instance()
+                .getCaptureDeleteLookupCondition(this);
+        if(potentialCaptureDeletes != null) {
+            for (Entry<Long, Set<Object>> entry : concourse
+                    .select(SECTION_KEY, potentialCaptureDeletes).entrySet()) {
+                long id = entry.getKey();
+                String __ = (String) Iterables.getLast(entry.getValue());
+                Class<? extends Record> clazz = Reflection.getClassCasted(__);
+                Record record = db.load(clazz, id);
+                Set<Field> fields = StaticAnalysis.instance()
+                        .getAnnotatedFields(record, CaptureDelete.class);
+                for (Field field : fields) {
+                    try {
+                        Object value = field.get(record);
+                        if(value instanceof Collection) {
+                            Collection<?> collection = (Collection<?>) value;
+                            while (collection.contains(this)) {
+                                collection.remove(this);
+                            }
+                        }
+                        else if(value.getClass().isArray()) {
+                            ArrayBuilder<Object> ab = ArrayBuilder.builder();
+                            Sequences.forEach(value, item -> {
+                                if(!item.equals(this)) {
+                                    ab.add(item);
+                                }
+                            });
+                            field.set(record, ab.build());
+                        }
+                        else if(value.equals(this)) {
+                            field.set(record, null);
+                        }
+                    }
+                    catch (ReflectiveOperationException e) {
+                        throw CheckedExceptions.wrapAsRuntimeException(e);
+                    }
+                }
+                record.saveWithinTransaction(concourse, new HashSet<>());
+            }
+        }
+
+        // Perform the deletion(s)
         concourse.clear(id);
+        for (Record record : waitingToBeDeleted) {
+            record.deleteWithinTransaction(concourse);
+        }
     }
 
     /**
@@ -2091,6 +2189,20 @@ public abstract class Record implements Comparable<Record> {
             }
         }
         return value;
+    }
+
+    /**
+     * Ensure that {@code record} is scheduled for
+     * {@link #deleteWithinTransaction(Concourse) deletion} alongside this
+     * {@link Record}.
+     * 
+     * @param record
+     */
+    private void ensureDeletion(Record record) {
+        if(!record.deleted) {
+            waitingToBeDeleted.add(record);
+            record.deleted = true;
+        }
     }
 
     /**
@@ -2500,8 +2612,7 @@ public abstract class Record implements Comparable<Record> {
 
         /**
          * A mapping from each {@link Record} class to each its non-internal
-         * keys,
-         * each of which is mapped to the associated {@link Field} object.
+         * keys, each of which is mapped to the associated {@link Field} object.
          */
         private final Map<Class<? extends Record>, Map<String, Field>> fieldsByClass;
 
@@ -2514,16 +2625,14 @@ public abstract class Record implements Comparable<Record> {
 
         /**
          * A collection containing each {@link Record} class that has at least
-         * one
-         * field whose type is a subclass of {@link Record}.
+         * one field whose type is a subclass of {@link Record}.
          */
         private final Set<Class<? extends Record>> hasRecordFieldTypeByClass;
 
         /**
          * A collection containing each {@link Record} class that itself or is
-         * the
-         * ancestors of a {@link Record} class that has at least one field whose
-         * type is a subclass of {@link Record}.
+         * the ancestors of a {@link Record} class that has at least one field
+         * whose type is a subclass of {@link Record}.
          */
         private final Set<Class<? extends Record>> hasRecordFieldTypeByClassHierarchy;
 
@@ -2532,6 +2641,38 @@ public abstract class Record implements Comparable<Record> {
          */
         private final Reflections reflection = new Reflections(
                 new SubTypesScanner());
+
+        /**
+         * A collection containing each {@link Record} class that has fields
+         * with annotations.
+         */
+        private final Map<Class<? extends Record>, Map<Class<? extends Annotation>, Set<Field>>> fieldAnnotationsByClass;
+
+        /**
+         * A mapping from each {@link Record} class to each of its related
+         * {@link Record} classes, each mapped to a set of field names that
+         * trigger deletion of records linked to it.
+         * <p>
+         * This structure enables efficient lookup of fields marked with the
+         * {@link JoinDelete} annotation, helping identify fields in linked
+         * records that should be deleted when the primary {@link Record} is
+         * deleted.
+         * </p>
+         */
+        private final Map<Class<? extends Record>, Map<Class<? extends Record>, Set<String>>> joinDeleteFieldsByClass;
+
+        /**
+         * A mapping from each {@link Record} class to each of its related
+         * {@link Record} classes, each mapped to a set of field names that
+         * trigger automatic reference nullification for records linked to it.
+         * <p>
+         * This structure enables efficient lookup of fields marked with the
+         * {@link CaptureDelete} annotation, helping identify fields in linked
+         * records that should be nullified when the primary {@link Record} is
+         * deleted.
+         * </p>
+         */
+        private final Map<Class<? extends Record>, Map<Class<? extends Record>, Set<String>>> captureDeleteFieldsByClass;
 
         /**
          * Construct a new instance.
@@ -2544,6 +2685,9 @@ public abstract class Record implements Comparable<Record> {
             this.fieldTypeArgumentsByClass = new HashMap<>();
             this.hasRecordFieldTypeByClass = new HashSet<>();
             this.hasRecordFieldTypeByClassHierarchy = new HashSet<>();
+            this.fieldAnnotationsByClass = new HashMap<>();
+            this.joinDeleteFieldsByClass = new HashMap<>();
+            this.captureDeleteFieldsByClass = new HashMap<>();
             Set<String> internalFieldNames = INTERNAL_FIELDS.keySet();
             reflection.getSubTypesOf(Record.class).forEach(type -> {
                 // Build class hierarchy
@@ -2568,6 +2712,15 @@ public abstract class Record implements Comparable<Record> {
                     if(Record.class.isAssignableFrom(field.getType())) {
                         hasRecordFieldTypeByClass.add(type);
                     }
+
+                    // Get annotations associated with each field
+                    for (Annotation annotation : field.getAnnotations()) {
+                        fieldAnnotationsByClass
+                                .computeIfAbsent(type, $ -> new HashMap<>())
+                                .computeIfAbsent(annotation.annotationType(),
+                                        $ -> new HashSet<>())
+                                .add(field);
+                    }
                 });
             });
 
@@ -2577,10 +2730,101 @@ public abstract class Record implements Comparable<Record> {
                 }
             });
 
+            // For each type, determine the types that have Link fields with the
+            // deletion hook annotations.
+            Map<Class<? extends Annotation>, Map<Class<? extends Record>, Map<Class<? extends Record>, Set<String>>>> hooks = ImmutableMap
+                    .of(JoinDelete.class, joinDeleteFieldsByClass,
+                            CaptureDelete.class, captureDeleteFieldsByClass);
+            hierarchies.keySet().forEach(type -> {
+                for (Class<? extends Record> incoming : hierarchies.keySet()) {
+                    hooks.forEach((annotation, data) -> {
+                        Set<Field> fields = getAnnotatedFields(incoming,
+                                annotation);
+                        for (Field field : fields) {
+                            if(isTypeCompatibleWithClassField(type, incoming,
+                                    field)) {
+                                data.computeIfAbsent(type, $ -> new HashMap<>())
+                                        .computeIfAbsent(incoming,
+                                                $$ -> new HashSet<>())
+                                        .add(field.getName());
+                            }
+                        }
+                    });
+                }
+            });
+
             // Now that the hierarchies and fields for each Record type have
             // been documented, go through again and compute the paths for each
             // one
             computeAllPossiblePaths();
+        }
+
+        /**
+         * Return a set of fields from a {@link Record} class that are annotated
+         * with a specific annotation.
+         * 
+         * @param <T> the type of record
+         * @param clazz the class of the record
+         * @param annotation the class of the annotation to look for
+         * @return a set of fields annotated with the specified annotation, or
+         *         an
+         *         empty set if none are found
+         * @throws IllegalArgumentException if the provided class is unknown
+         */
+        public <T extends Record> Set<Field> getAnnotatedFields(Class<T> clazz,
+                Class<? extends Annotation> annotation) {
+            try {
+                return fieldAnnotationsByClass
+                        .getOrDefault(clazz, ImmutableMap.of())
+                        .getOrDefault(annotation, ImmutableSet.of());
+            }
+            catch (NullPointerException e) {
+                throw new IllegalArgumentException(
+                        "Unknown Record type: " + clazz);
+            }
+        }
+
+        /**
+         * Return a set of fields from an instance of a {@link Record} that are
+         * annotated with a specific annotation.
+         * 
+         * @param <T> the type of record
+         * @param record the record instance
+         * @param annotation the class of the annotation to look for
+         * @return a set of fields annotated with the specified annotation, or
+         *         an
+         *         empty set if none are found
+         * @throws IllegalArgumentException if the provided record type is
+         *             unknown
+         */
+        public <T extends Record> Set<Field> getAnnotatedFields(T record,
+                Class<? extends Annotation> annotation) {
+            return getAnnotatedFields(record.getClass(), annotation);
+        }
+
+        /**
+         * Return a {@link Criteria} instance that defines the condition to
+         * locate records linked to the specified {@link Record} that should
+         * have fields nullified upon deletion of the specified record.
+         * <p>
+         * This condition is constructed by identifying fields in other
+         * {@link Record} classes that are annotated with {@link CaptureDelete}
+         * and linked to the specified {@code record}. These fields are grouped
+         * by related classes, and then a combined criteria is formed to match
+         * against the links found.
+         * </p>
+         *
+         * @param record the {@link Record} instance for which the nullification
+         *            condition is being determined
+         * @return a {@link Criteria} object representing the lookup condition
+         *         for records linked to the provided {@code record} that should
+         *         have references nullified, or {@code null} if no such links
+         *         are found
+         */
+        public <T extends Record> Criteria getCaptureDeleteLookupCondition(
+                Record record) {
+            return getDeletionHookLookupCondition(record,
+                    captureDeleteFieldsByClass);
         }
 
         /**
@@ -2648,6 +2892,30 @@ public abstract class Record implements Comparable<Record> {
          */
         public <T extends Record> Collection<Field> getFields(T record) {
             return getFields(record.getClass());
+        }
+
+        /**
+         * Return a {@link Criteria} instance that defines the condition to
+         * locate records linked to the specified {@link Record} for deletion.
+         * <p>
+         * This condition is constructed by identifying fields in other
+         * {@link Record} classes that are annotated with {@link JoinDelete} and
+         * linked to the specified {@code record}. These fields are grouped by
+         * related classes, and then a combined criteria is formed to match
+         * against the links found.
+         * </p>
+         *
+         * @param record the {@link Record} instance for which the delete
+         *            condition is being determined
+         * @return a {@link Criteria} object representing the delete lookup
+         *         condition for records linked to the provided {@code record};
+         *         or {@code null} if no such links are found
+         */
+        @Nullable
+        public <T extends Record> Criteria getJoinDeleteLookupCondition(
+                Record record) {
+            return getDeletionHookLookupCondition(record,
+                    joinDeleteFieldsByClass);
         }
 
         /**
@@ -2763,6 +3031,102 @@ public abstract class Record implements Comparable<Record> {
                 pathsByClassHierarchy.put(type, computePathsHierarchy(type,
                         hierarchies, fieldsByClass));
             });
+        }
+
+        /**
+         * Return a {@link Criteria} instance that defines the condition to
+         * locate records linked to the specified {@link Record} based on a
+         * deletion hook.
+         * This method is designed to support various deletion hooks, such as
+         * {@link JoinDelete} and {@link CaptureDelete}, by using the provided
+         * data map to determine which fields in other records should be
+         * affected when the specified {@code record} is deleted.
+         * <p>
+         * This condition is constructed by gathering all fields in related
+         * {@link Record} classes that are marked with the specified deletion
+         * annotation and link to the specified {@code record}. Each associated
+         * class and its fields are combined into a single {@link Criteria}
+         * expression that matches records with fields needing to be either
+         * deleted or nullified, depending on the deletion hook type.
+         * </p>
+         *
+         * @param record the {@link Record} instance for which the deletion
+         *            condition is being constructed
+         * @param data a mapping that defines the deletion hooks for each
+         *            {@link Record} class, where each key is a {@link Record}
+         *            type linked by the hook, and each value is a set of field
+         *            names annotated with the deletion hook
+         * @return a {@link Criteria} instance representing the lookup condition
+         *         for records linked to the provided {@code record}, or
+         *         {@code null} if no linked records with the deletion hook are
+         *         found
+         */
+        @Nullable
+        private <T extends Record> Criteria getDeletionHookLookupCondition(
+                Record record,
+                Map<Class<? extends Record>, Map<Class<? extends Record>, Set<String>>> data) {
+            Criteria condition = null;
+            Map<Class<? extends Record>, Set<String>> components = data
+                    .getOrDefault(record.getClass(), ImmutableMap.of());
+            for (Entry<Class<? extends Record>, Set<String>> entry : components
+                    .entrySet()) {
+                Class<? extends Record> type = entry.getKey();
+                Set<String> keys = entry.getValue();
+                Criteria typePart = Criteria.where().key(SECTION_KEY)
+                        .operator(Operator.EQUALS).value(type.getName());
+                Criteria linkPart = null;
+                for (String key : keys) {
+                    Criteria $ = Criteria.where().key(key)
+                            .operator(Operator.LINKS_TO).value(record.id);
+                    if(linkPart == null) {
+                        linkPart = $;
+                    }
+                    else {
+                        linkPart = Criteria.where().group(linkPart).or()
+                                .group($);
+                    }
+                }
+                Criteria part = Criteria.where().group(typePart).and()
+                        .group(linkPart);
+                if(condition == null) {
+                    condition = part;
+                }
+                else {
+                    condition = Criteria.where().group(condition).or()
+                            .group(part);
+                }
+            }
+            return condition;
+        }
+
+        /**
+         * Determine whether the specified {@code type} is compatible with or
+         * assignable to the given {@code field} within {@code clazz}. This
+         * check includes both direct assignment compatibility as well as
+         * compatibility with the field's type arguments.
+         * 
+         * @param type the {@link Record} type to check for compatibility with
+         *            the field
+         * @param clazz the class that contains the {@code field}
+         * @param field the {@link Field} in {@code clazz} to check against
+         * @return {@code true} if {@code type} is compatible with the field, or
+         *         {@code false} otherwise
+         */
+        private boolean isTypeCompatibleWithClassField(
+                Class<? extends Record> type, Class<? extends Record> clazz,
+                Field field) {
+            if(field.getType().isAssignableFrom(type)) {
+                return true;
+            }
+            else {
+                for (Class<?> typeArg : getTypeArguments(clazz,
+                        field.getName())) {
+                    if(typeArg.isAssignableFrom(type)) {
+                        return true;
+                    }
+                }
+            }
+            return false;
         }
 
     }